Rename join package to elastic
